fe_updated User's Dashboard. added loading icon for each category
