+++ frontend/src/components/User/Sidebar.jsx
@@ -20,7 +20,7 @@ function Sidebar({ currentUserRoute }) {
   }
 
   return (
-    <aside className="cust-bg-background sticky left-0 top-0 z-30 whitespace-nowrap border-r border-gray-300 md:w-48">
+    <aside className="sticky left-0 top-0 z-30 whitespace-nowrap border-r border-gray-300 md:w-48">
       <div className="px-2 py-3">
         <ul className="space-y-2 font-medium">
           <li>
